python-backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
joblib==1.3.2
numpy==1.26.4
pydantic==2.6.1 